+++ lib/modules/ipfs/index.js
@@ -12,26 +12,48 @@ class IPFS {
     this.events = embark.events;
     this.buildDir = options.buildDir;
     this.storageConfig = embark.config.storageConfig;
+    this.namesystemConfig = embark.config.namesystemConfig;
     this.embark = embark;
 
     this.webServerConfig = embark.config.webServerConfig;
     this.blockchainConfig = embark.config.blockchainConfig;
 
-    if (!this.isIpfsEnabledInTheConfig()) {
-      return;
+    if (this.isIpfsEnabledInTheConfig() || this.isIpnsEnabledInTheConfig()) {
+      this.downloadIpfsApi();
+      this.addDefaultToEmbarkJS();
     }
 
-    this.setServiceCheck();
-    this.addProviderToEmbarkJS();
-    this.addObjectToConsole();
-    this.registerUploadCommand();
+    if (this.isIpfsEnabledInTheConfig()) {
+      this.setServiceCheck();
+      this.addStorageProviderToEmbarkJS();
+      this.addObjectToConsole();
+      this.registerUploadCommand();
 
-    this._checkService((err) => {
-      if (!err) {
-        return;
+      this._checkService((err) => {
+        if (!err) {
+          return;
+        }
+        self.logger.info("IPFS node not found, attempting to start own node");
+        self.startProcess(() => {});
+      });
+    }
+
+    if (this.isIpnsEnabledInTheConfig()) {
+      this.addNamesystemProviderToEmbarkJS();
+      this.setNamesystemProvider();
+    }
+  }
+
+  downloadIpfsApi() {
+    const self = this;
+
+    self.events.request("version:get:ipfs-api", function(ipfsApiVersion) {
+      let currentIpfsApiVersion = require('../../../package.json').dependencies["ipfs-api"];
+      if (ipfsApiVersion !== currentIpfsApiVersion) {
+        self.events.request("version:getPackageLocation", "ipfs-api", ipfsApiVersion, function(err, location) {
+          self.embark.registerImportFile("ipfs-api", fs.dappPath(location));
+        });
       }
-      self.logger.info("IPFS node not found, attempting to start own node");
-      self.startProcess(() => {});
     });
   }
 
@@ -79,21 +101,24 @@ class IPFS {
     utils.getJson(url, cb);
   }
 
-  addProviderToEmbarkJS() {
-    const self = this;
-
-    self.events.request("version:get:ipfs-api", function(ipfsApiVersion) {
-      let currentIpfsApiVersion = require('../../../package.json').dependencies["ipfs-api"];
-      if (ipfsApiVersion !== currentIpfsApiVersion) {
-        self.events.request("version:getPackageLocation", "ipfs-api", ipfsApiVersion, function(err, location) {
-          self.embark.registerImportFile("ipfs-api", fs.dappPath(location));
-        });
-      }
-    });
-
+  addDefaultToEmbarkJS() {
     let code = "";
-    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs.js')).toString();
+    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs' , 'default.js')).toString();
+
+    this.embark.addCodeToEmbarkJS(code);
+  }
+
+  addStorageProviderToEmbarkJS() {
+    let code = "";
+    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs', 'storage.js')).toString();
     code += "\nEmbarkJS.Storage.registerProvider('ipfs', __embarkIPFS);";
+
+    this.embark.addCodeToEmbarkJS(code);
+  }
+
+  addNamesystemProviderToEmbarkJS() {
+    let code = "";
+    code += "\n" + fs.readFileSync(utils.joinPath(__dirname, 'embarkjs', 'name.js')).toString();
     code += "\nEmbarkJS.Names.registerProvider('ipns', __embarkIPFS);";
 
     this.embark.addCodeToEmbarkJS(code);
@@ -134,6 +159,17 @@ class IPFS {
     let {enabled, available_providers, dappConnection} = this.storageConfig;
     return enabled && (available_providers.indexOf('ipfs') > 0 || dappConnection.find(c => c.provider === 'ipfs'));
   }
+
+  isIpnsEnabledInTheConfig() {
+    let {enabled, available_providers, provider} = this.namesystemConfig;
+    return enabled && available_providers.indexOf('ipns') > 0 && provider === 'ipns';
+  }
+
+  setNamesystemProvider() {
+    let code = `\nEmbarkJS.Names.setProvider('ipns', ${JSON.stringify(this.storageConfig.dappConnection[0] || {})});`;
+    this.embark.addProviderInit('names', code, this.isIpnsEnabledInTheConfig.bind(this));
+  }
+
 
 }
